// supabase/functions/_shared/models/Copilot.ts
export interface CopilotModel {
  id?: string
  uid?: string
  /**
   * name
   */
  name: string
  /**
   * description
   */
  description: string
  /**
   * prompt
   */
  prompt?: string
  /**
   * interactions
   */
  interactions?: number
  /**
   * visibility
   */
  visibility?: 'public' | 'private'
  /**
   * 创建时间
   */
  createdAt?: Date
  /**
   * 修改时间
   */
  updatedAt?: Date
}